// src/api/portal-api/models/book-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Prueba API
 * This Api will be responsible for overall data distribution
 *
 * The version of the OpenAPI document: v1
 * Contact: laihan758.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface BookDto
 */
export interface BookDto {
    /**
     * 
     * @type {number}
     * @memberof BookDto
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof BookDto
     */
    'title'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BookDto
     */
    'description'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof BookDto
     */
    'pageCount'?: number;
    /**
     * 
     * @type {string}
     * @memberof BookDto
     */
    'excerpt'?: string | null;
}